Stop importing the native node 'path' module in shared/paths-util.js by moving it to app/main
